% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{adf}
\alias{adf}
\title{Augmented Dickey-Fuller (ADF) test}
\usage{
adf(d, vars = colnames(d), type = c("none", "drift", "trend"), lags = 1,
  filename = NULL, digits = 3, verbose = TRUE, ...)
}
\arguments{
\item{d}{Data frame containing the time series in column-wise format.}

\item{vars}{Column names to check for stationarity. If not specified,
then all columns are tested.}

\item{type}{Test type, either \code{"none"}, \code{"drift"} or
\code{"trend"}. See \code{\link[urca]{ur.df}} for details.}

\item{lags}{Number of lags for endogenous variable to be included.
Default is 1. See \code{\link[urca]{ur.df}} for details.}

\item{filename}{Optional filename to export the table as LaTeX.
Default is \code{NULL}, i.e. no export.}

\item{digits}{Number of digits to be printed (default: 3).}

\item{verbose}{Flag if the result of each ADF test is printed.
Default is yes (\code{TRUE}).}

\item{...}{Further parameters passed on to \code{\link[urca]{ur.df}}.}
}
\value{
P-values from ADF test in a simple table. If P-values are
< 0.05, this is an indication of stationarity.
}
\description{
Performs the Augmented Dickey-Fuller (ADF) test to check if a time
series is stationary. The result comes in a pretty format.
}
\examples{
data(USArrests)
adf(USArrests)
adf(USArrests, vars = c("Murder", "Rape"), type = "drift",
    filename = "adf.tex", verbose = FALSE)
unlink("adf.tex")
}

