% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{allDigitsLabels}
\alias{allDigitsLabels}
\title{ggplot labels showing all digits}
\usage{
allDigitsLabels(l)
}
\arguments{
\item{l}{Labels}
}
\value{
Labels in the desired format
}
\description{
Allows to adjust the numbers display on both axes for ggplot2 plots. The format is such that all digits
are displayed.
}
\examples{
library(ggplot2)
df <- data.frame(x=rnorm(100)/1000, y=rnorm(100)/1000)
ggplot(df, aes(x=x, y=y)) +
  geom_point() +
  scale_x_continuous(labels=allDigitsLabels) +
  scale_y_continuous(labels=allDigitsLabels)
}

