% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{ceil}
\alias{ceil}
\title{Downward rounding of numbers}
\usage{
ceil(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with integers after rounding downwards to the next integer.
}
\description{
Takes a single numeric value or a vector and returns the smallest integers not less
than the corresponding elements.
}
\details{
A simple wrapper to \code{ceiling} for a more consistent naming convention
with other programming languages.
}
\examples{
ceil(3.14)
ceil(c(-3.2, 4, 0, 7.6, 7.1))
}

