% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{cointegrationTable}
\alias{cointegrationTable}
\title{Cointegration test}
\usage{
cointegrationTable(d, vars = colnames(d), type = c("eigen", "trace"),
  K = NULL, filename = paste0("cointegration_", type[1], ".tex"),
  digits = 3, ...)
}
\arguments{
\item{d}{Data frame containing the time series in column-wise format.}

\item{vars}{Column names to check for stationarity. If not specified,
then all columns are tested.}

\item{type}{The test to be conducted, either \code{"eigen"} or
\code{"trace"}. See \code{\link[urca]{ur.df}} for details.}

\item{K}{The lag order of the series (levels) in the VAR.}

\item{filename}{Filename to export the table as LaTeX.}

\item{digits}{Number of digits to be printed (default: 3).}

\item{...}{Further parameters passed on to \code{\link[urca]{ca.jo}}.}
}
\value{
Table with result from cointegration test.
}
\description{
Conducts a cointegration test based on the Johansen procedure for
multivariate time series.
}
\examples{
data(USArrests)
cointegrationTable(USArrests, vars = c("Murder", "Rape"), K = 2)
unlink("cointegration_eigen.tex")
}

