% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{completeLowResolutionData}
\alias{completeLowResolutionData}
\title{Duplicates data with low time resolution to match high resolution}
\usage{
completeLowResolutionData(date_target, df, by_column = "Date")
}
\arguments{
\item{date_target}{Time series of desired output (high resolution).}

\item{df}{Data frame with the low resolution data.}

\item{by_column}{String which identifies the date column in \code{df}.
As a default, it uses \code{"Date"}.}
}
\value{
Returns data frame with copied entries. The response is sorted
in chronological order.
}
\description{
Let data in low resolution be given (e.g. monthly, quarterly). This
is then matched to a high resolution (e.g. daily) and the corresponding
entries are copied.
}
\examples{
ts <- data.frame(Date = seq(from = as.Date("2000-01-01"), to = as.Date("2000-03-31"), by = "1 day"))

df_monthly <- data.frame(Month = c(as.Date("2000-01-31"),
                                   as.Date("2000-02-29"),
                                   as.Date("2000-03-31")),
                         Values = 1:3)
df_daily <- completeLowResolutionData(ts$Date, df_monthly, "Month")

# example of how to bind things together
library(dplyr)
ts <- ts \%>\%
  left_join(df_daily, by = c("Date" = "Month"))
}

