% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsd}
\alias{cumadev}
\alias{cumkurtosis}
\alias{cumsd}
\alias{cumskewness}
\title{Cumulativate versions of skewness, kurtosis, sd, adv}
\usage{
cumsd(x)

cumskewness(x)

cumkurtosis(x)

cumadev(x)
}
\arguments{
\item{x}{For \code{cumskewness}, \code{cumkurtosis}, \code{cumsd},
\code{cumadv} an integer or numeric vector}
}
\description{
ResearchGroupTools adds \code{cumskewness}, \code{cumkurtosis},
\code{cumadv}, and \code{cumsd} to complete
R's set of cumulate functions to match the aggregation functions available
in most databases
}

