% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptiveStatistics}
\alias{descriptiveStatistics}
\title{Pretty summary statistics}
\usage{
descriptiveStatistics(x, digits = 3, subset = c("mean", "median", "min",
  "max", "sd", "skew", "kurtosis"), filename = "table_descriptives.tex")
}
\arguments{
\item{x}{Dataset in the form of a matrix or a \code{data.frame}.}

\item{digits}{Number of digits to be printed (default: 3).}

\item{subset}{Variables to be selected.}

\item{filename}{Path to which a LaTeX copy is exported (default is
\code{table_descriptives.tex}). If \code{NULL}, no file is generated.}
}
\value{
Matrix with pretty output.
}
\description{
Calculates summary statistics for dataset and display key variables in a nice
format. It also writes the table into a LaTeX file if desired.
}
\examples{
data(USArrests)
descriptiveStatistics(USArrests)
unlink("table_descriptives.tex")
}

