% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{differences}
\alias{differences}
\title{Lagged differences}
\usage{
differences(x, lag = 1, order = 1, na_padding = TRUE)
}
\arguments{
\item{x}{A numeric vector containing the values for which returns are calculated.}

\item{lag}{Integer indicating which lag to use (default: 1).}

\item{order}{Order of differences (default: 1).}

\item{na_padding}{Flag whether to pad truncated values with NA's (default: TRUE).}
}
\value{
Numeric vector with differences of specified lag and order.
}
\description{
Calculates differences between values with specified lag.
}
\note{
Parameter \code{na_padding} behaves slightly different from \code{\link{diff}}
in order to achieve consistency with \code{\link[dplyr]{lead}} and \code{\link[dplyr]{lag}}.
}
\examples{
differences(1:10)
differences(c(1, 2, 4, 8, 16, 32))
differences(c(1, 2, 4, 8, 16, 32), order = 2)
differences(c(1, 2, 4, 8, 16, 32), na_padding = FALSE)
}
\seealso{
\code{\link{diff}}, \code{\link{returns}}
}

