% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Exports object to LaTeX}
\usage{
export(object, filename = NULL, digits = attr(object, "digits"))
}
\arguments{
\item{object}{A return value from e.g. \code{\link{descriptiveStatistics}}.}

\item{filename}{String specifying the filname for export. If \code{NULL}, the
routine uses the name based on the class of \code{object}.}

\item{digits}{Number of digits. If not specified otherwise, the number of
digits is directly infered from the attribute of \code{object}.}
}
\description{
Export of objects from this package to LaTeX.
}

