% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{exportAdfDifferences}
\alias{exportAdfDifferences}
\title{Export multiple ADF tests}
\usage{
exportAdfDifferences(adf_levels, adf_diff1, filename = "adf.tex",
  digits = 3)
}
\arguments{
\item{adf_levels}{ADF test for variables in levels as returned by \code{\link{adf}}.}

\item{adf_diff1}{ADF test for variables in differences as returned by \code{\link{adf}}.}

\item{filename}{LaTeX file to export to. Defaul name is \code{adf.tex}.}

\item{digits}{Number of digits in the resulting table. Default is 3.}
}
\value{
Pretty table for output in R.
}
\description{
Exports augmented Dickey-Fuller (ADF) tests of variables in levels and variables in
differences into a single LaTeX file. It keeps the original order of the variables,
but always displays the results for the levels first, followed by the differences and
then continues with the next variables.
}
\examples{
data(USArrests)
adf_levels <- adf(USArrests)
adf_diff1 <- adf(data.frame(Murder = diff(USArrests$Murder),
                            Assault = diff(USArrests$Assault),
                            UrbanPop = diff(USArrests$UrbanPop),
                            Rape = diff(USArrests$Rape)))
exportAdfDifferences(adf_levels, adf_diff1)
unlink("adf.tex")
}

