% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extract_spikeslab}
\alias{extract_spikeslab}
\title{Function to perform \code{textreg} export of \code{spikeslab} objects.}
\usage{
extract_spikeslab(model, standardized = TRUE, excludeNoise = TRUE, ...)
}
\arguments{
\item{model}{Object of type \code{spikeslab}.}

\item{...}{Further parameters (currently ignore for \code{spikeslab}).}

\item{standarized}{A logical value. If \code{TRUE}, standardized coefficient are used (default).}

\item{exludeNoise}{A logical value. If \code{TRUE}, non-relevant regressors are excluded (default).}
}
\value{
Object of type \code{\link[texreg]{texreg}}.
}
\description{
Function to perform \code{textreg} export of \code{spikeslab} objects.
}

