% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extract_tvalues}
\alias{extract_tvalues}
\title{Extract t-values from lm object}
\usage{
extract_tvalues(model, hide = NULL)
}
\arguments{
\item{model}{Object of type \code{lm}.}

\item{hide}{A string. All variables starting with that name are excluded. If \code{NULL} (default),
no variables are omitted.}
}
\value{
Vector of type \code{numeric}.
}
\description{
Function extracts the t-values from an \code{lm} object as given by the \code{summary} function.
}
\examples{
\dontrun{
x1 <- 1:100
x2 <- rep(c(1, 2), 50)
y <- x1 + x2 + rnorm(100)

m <- lm(y ~ x1 + x2)
extract_tvalues(m)
}
}

