% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extractRegressionStatistics}
\alias{extractRegressionStatistics}
\title{Extracts key statistics of regression}
\usage{
extractRegressionStatistics(model)
}
\arguments{
\item{model}{Object of type \code{lm}.}
}
\value{
Returns data frame with named columns.
}
\description{
Function extracts various parameter from an \code{lm} object, including the (adjuste) R-squared,
AIC, BIC and the outcome of the F-test.
}
\examples{
x <- 1:10
y <- 1 + x + rnorm(10)
m <- lm(y ~ x)

extractRegressionStatistics(m)
}

