% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{findColsNA}
\alias{findColsNA}
\title{Names of columns with \code{NA} values}
\usage{
findColsNA(d)
}
\arguments{
\item{d}{A matrix or \code{data.frame}.}
}
\value{
Vector with column identifiers
}
\description{
Identifies columns that contain an \code{NA} value
}
\examples{
m <- matrix(letters[c(1, 2, NA, 3, NA, 4, 5, 6, 7, 8)],
            ncol = 2, byrow = FALSE)
colnames(m) <- c("x", "y")
m
findColsNA(m)
showColsNA(m)

d <- data.frame(x = 1:10)
d
findColsNA(d)
showColsNA(d)
}
\seealso{
\code{\link{showColsNA}}
}

