% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getRowsOutlierRemoval}
\alias{getRowsOutlierRemoval}
\title{Identify row numbers for outlier removal}
\usage{
getRowsOutlierRemoval(model, cutoff = 0.5)
}
\arguments{
\item{model}{Object of type \code{lm}.}

\item{cutoff}{Relative cutoff on each side in percent (default: \code{0.5}, i.e. 0.5\% at each end).}
}
\value{
Returns vector with indices of the observations to be removed.
}
\description{
Function removes outliers at a relative percentage at both ends.
}
\examples{
d <- data.frame(x = 1:200, y = 1:200 + rnorm(200))
m <- lm(y ~ x, d)

idx_rm <- getRowsOutlierRemoval(m)

m <- lm(y ~ x, d[-idx_rm, ])
}

