% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%+\%}
\alias{\%+\%}
\title{Operator for concatenating}
\usage{
... \%+\% NA
}
\arguments{
\item{...}{Individual objects that are to be concatenated.}
}
\value{
String with concatenated objects.
}
\description{
Operator is a handy alternative to \code{paste()}, which allows you
to concatenate strings.
}
\examples{
"a" \%+\% "b"
3 \%+\% 4
`\%+\%`(letters)
}

