% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{impulseResponsePlot}
\alias{impulseResponsePlot}
\title{Calculation and pretty plot of impulse response function}
\usage{
impulseResponsePlot(var, impulse, response, ...)
}
\arguments{
\item{var}{Object package with impulse response function, i.e. type
\code{varest} from the \code{vars}}

\item{impulse}{String identifier which variable experiences a shock.}

\item{response}{String identfier which variable is the response.}

\item{...}{Further arguments passed on to \code{\link{plotIrf}}}
}
\value{
Object of \code{\link[ggplot2]{ggplot}}.
}
\description{
Computes and plots the impulse response function in black/white manner for
suitability with common journals.
}
\examples{
library(vars)
data(Canada)
# For VAR
var.2c <- VAR(Canada, p = 2, type = "const")
\dontrun{
impulseResponsePlot(var.2c, impulse = "e", response = "prod", ylab = "Production")
}
}

