% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{jplot}
\alias{jplot}
\title{Plotting routine in journal-style}
\usage{
jplot(data = NULL, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{Default dataset used for plotting.}

\item{mapping}{Default list of aesthetics.}

\item{...}{Further arguments passed to \code{\link[ggplot2]{ggplot}}.}

\item{environment}{Alternative search space if a varaible is not found in
\code{data}.}
}
\value{
Returns a masked \code{ggplot} object that is encoded as class
"\code{RGT_plot}".
}
\description{
Wrapper to \code{ggplot} with automated change to gray-scale coloring.
}
\examples{
library(ggplot2)

df <- data.frame(x = 1:20,
                 y = 1:20,
                 z = as.factor(rep(1:4, each = 5)))

ggplot(df) +
  geom_line(aes(x = x, y = y, color = z))

jplot(df) +
  geom_line(aes(x = x, y = y, color = z))

ggplot(df) +
  geom_point(aes(x = x, y = y, color = z))

jplot(df) +
  geom_point(aes(x = x, y = y, color = z))
}

