% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{lags}
\alias{lags}
\title{Multiple lags of time series}
\usage{
lags(x, k, ...)
}
\arguments{
\item{x}{A vector or univariate time series.}

\item{k}{A numeric value or a vector denoting which lags should be
retrieved.}

\item{...}{Further parameters passed to \code{\link[dplyr]{lag}}.}
}
\value{
Matrix where the number of rows is identical to the length
of \code{x}. The number of columns is identical the number of entries
in \code{k}.
}
\description{
Multiple lags of time series
}
\examples{
x <- 1:10
lags(x, c(1, 2, 3))

z <- ts(1:10)
lags(z, c(1, 2, 5))
}

