% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{last_non_NA}
\alias{last_non_NA}
\title{Returns the last non-NA entry}
\usage{
last_non_NA(d)
}
\arguments{
\item{d}{Vector with numeric entries and possibly \code{NA} values.}
}
\value{
Last value in \code{d} with non-NA entry.
}
\description{
Finds the last value in a vector which is not \code{NA} and
returns it. If all entries are \code{NA}, it returns an \code{NA}
}
\examples{
last_non_NA(c(1, 2, 3, 4, NA))

values <- 1:100
values[sample(1:100, 10)] <- NA
df <- cbind(Year = c(rep(2000, 5), rep(2001, 5)),
            as.data.frame(matrix(values, nrow = 10)))

library(dplyr)
df \%>\%
  group_by(Year) \%>\%
  summarize_each(funs(last_non_NA)) \%>\%
  ungroup() \%>\%
  head()
}

