% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{Library}
\alias{Library}
\title{Wrapper function for loading and installing packages}
\usage{
Library(packages, ...)
}
\arguments{
\item{packages}{String or vector of strings identifying the packages.}

\item{...}{Strings with further package names}
}
\description{
Installs package if these are not yet available on the current machine. Afterwards,
all packages are loaded.
}
\examples{
\dontrun{
Library("lmtest")
Library("lmtest", "psych")
}
}

