% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{linePlot}
\alias{linePlot}
\title{Rapid line plot with ggplot2}
\usage{
linePlot(x, y = NULL)
}
\arguments{
\item{x}{Values on x-axis if a vector. If it contains two columns, that it plots the first column on the x-axis, the second on the y-axis.}

\item{y}{Optional values on the y-axis.}
}
\value{
Object of type \code{\link[ggplot2]{ggplot}}.
}
\description{
Creates simple line plot with ggplot2.
}
\examples{
linePlot(1:10)

x <- seq(0, 4, length.out = 100)
linePlot(x, sin(x))
}

