% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{loadRegressionLibraries}
\alias{loadRegressionLibraries}
\title{Loads and installs default libraries for regressions}
\usage{
loadRegressionLibraries()
}
\description{
Bundles routines for regressions as the mapping between test routine
and originating package is not always obvious.
}
\examples{
\dontrun{
  loadRegressionLibraries()
}
}

