% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{makeFormula}
\alias{makeFormula}
\title{Create formula from strings}
\usage{
makeFormula(dependent, independent, dummies = NULL)
}
\arguments{
\item{dependent}{Name of dependent variable.}

\item{independent}{String or vector of strings with independent variables.}

\item{dummies}{Optional name for dummies. If \code{NULL} (default), then dummies are omitted.}
}
\description{
Create formula from strings
}
\examples{
makeFormula("y", "x")
makeFormula("y", c("x1", "x2", "x3"))
makeFormula("y", c("x1", "x2", "x3"), "dummies")
}

