% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{plotImpulseResponseFunction}
\alias{plotImpulseResponseFunction}
\title{Pretty plot of impulse response function}
\usage{
plotImpulseResponseFunction(irf, ylab)
}
\arguments{
\item{irf}{Object of type \code{\link[vars]{varirf}} with impulse response function.}

\item{ylab}{ylab}
}
\description{
Plots the impulse response function in black/white manner for
suitability with common journals.
}

