% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print.RGT_plot}
\alias{print.RGT_plot}
\title{Routine to call plotting functionality}
\usage{
\method{print}{RGT_plot}(x, ...)
}
\arguments{
\item{x}{Object of type \code{RGT_plot}.}

\item{...}{Further arguemnts passed on to \code{ggplot2}.}
}
\description{
Function call is redirected to \code{ggplot2} in order to plot the argument.
}
\examples{
library(ggplot2)

df <- data.frame(x = 1:20,
                 y = 1:20,
                 z = as.factor(rep(1:4, each = 5)))

ggplot(df) +
  geom_line(aes(x = x, y = y, color = z))

jplot(df) +
  geom_line(aes(x = x, y = y, color = z))

ggplot(df) +
  geom_point(aes(x = x, y = y, color = z))

jplot(df) +
  geom_point(aes(x = x, y = y, color = z))
}

