% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{pull}
\alias{pull}
\alias{pull_ith}
\alias{pull_string}
\title{Extracts a column from a tbl object}
\source{
Tommy O'Dell: \url{http://stackoverflow.com/questions/21618423/extract-a-dplyr-tbl-column-as-a-vector}
}
\usage{
pull(object, column)

pull_string(object, column)

pull_ith(object, column)
}
\arguments{
\item{object}{Object of type \code{\link[dplyr]{tbl}}}

\item{column}{Column identifier, name as string or integer depending on function.}
}
\value{
Column as vector
}
\description{
Extracts a column from a tbl object
}
\examples{
library(dplyr)

d <- data_frame(x = 1:10,
                y = rnorm(10))
d \%>\% pull(x)
d \%>\% pull("x")

v <- "x"
d \%>\% pull_string(v)

d \%>\% pull_ith(1)
}

