% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{rebuildPackage}
\alias{rebuildPackage}
\title{Builds, loads and checks package}
\usage{
rebuildPackage(updateReadme = FALSE)
}
\arguments{
\item{updateReadme}{Flag whether to update the README.Rmd (default: FALSE).}
}
\description{
Simple wrapper to rebuild, load and check package. This function specifically
updates the manual as well.
}
\examples{
\dontrun{
  rebuildPackage()
}
}

