% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regressionStepwise}
\alias{regressionStepwise}
\title{Step-wise regression}
\usage{
regressionStepwise(formula, data = NULL, subset = NULL, dummies = NULL,
  cutoff = NULL, rmDummyNA = TRUE, vcov = NULL)
}
\arguments{
\item{formula}{of type \code{formula}.}

\item{data}{An optional data frame contain the variables in the model (excluding the
dummy variables).}

\item{subset}{Vector of integers or booleans defining the subset of observations to be
used.}

\item{dummies}{String denoting the name of the variable (i.e. matrix or data frame)
containing all dummy variables.}

\item{cutoff}{Relative cutoff on each side in percent (default: \code{NULL}). Values
are given in percient, e.g. \code{0.5} represents 0.5\%
at each end).}

\item{rmDummyNA}{Boolean indicating whether to remove dummy variables with NA
coefficient (default: removal).}

\item{vcov}{Estimator used for computing the covariance matrix. Default is \code{NULL}
which results in to the ordinary least squares estimator. Alternatives, are, for
instance, \code{NeweyWest} from the \code{sandwich} package.}
}
\value{
Returns a list with different, estimated models.
}
\description{
Function performs default regression via ordinary least squares in step-wise fashion.
That is, it iteratively includes an additional regressor one-by-one. It also supports dummy
variables which are not included in the dataset \code{data}, but in a global variable
attached to a formula. With this input, this function can filter for a subset, remove
outliers at a certain cutoff and remove dummies that are NA.
}
\note{
The dummy term is included in all regressions (if present).
}
\examples{
x1 <- 1:100
x2 <- sin(1:100)
clusters <- rep(c(1, 2), 50)
dummies <- model.matrix(~ clusters)
y <- x1 + x2 + clusters + rnorm(100)
d <- data.frame(x1 = x1, x2 = x2, y = y)

models <- regressionStepwise(formula("y ~ x1 + x2 + dummies"), data = d, subset = 1:90,
                             dummies = "dummies", cutoff = 0.5)

length(models)

library(texreg)
texreg(models, omit.coef = "dummies")

library(sandwich)

models <- regressionStepwise(formula("y ~ x1 + x2 + dummies"), data = d, subset = 1:90,
                             dummies = "dummies", cutoff = 0.5, vcov = NeweyWest)
texreg(models, omit.coef = "dummies")
}

