% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{removeOutlierObservations}
\alias{removeOutlierObservations}
\title{Removes observations with outliers}
\usage{
removeOutlierObservations(df, variables = colnames(df), cutoff = 0.5)
}
\arguments{
\item{df}{Data frame with input variables}

\item{variables}{Variables used for trimming. By default, all variables are
included.}

\item{cutoff}{Relative cutoff on each side in percent (default: \code{0.5},
i.e. 0.5\% at each end).}
}
\value{
Returns vector with indices of the observations to be removed.
}
\description{
Function removes outliers for a given set of variables at a relative
percentage at both ends.
}
\examples{
d <- data.frame(x1 = rnorm(200), x2 = rnorm(200), y = rnorm(200))

d_trimmed <- removeOutlierObservations(d)
dim(d_trimmed)

d_trimmed <- removeOutlierObservations(d, variables = "y", cutoff = 2.0)
dim(d_trimmed)

d_trimmed <- removeOutlierObservations(d, variables = c("x1", "x2"), cutoff = 2.0)
dim(d_trimmed)
}

