% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{returns}
\alias{returns}
\title{Lagged returns}
\usage{
returns(x, lag = 1, na_padding = TRUE)
}
\arguments{
\item{x}{A numeric vector containing the values for which returns are calculated.}

\item{lag}{Integer indicating which lag to use (default: 1).}

\item{na_padding}{Flag whether to pad truncated values with NA's (default: TRUE).}
}
\value{
Numeric vector with returns of specified lag.
}
\description{
Calculates returns between values with specified lag.
}
\note{
Parameter \code{na_padding} behaves slightly different from \code{\link{diff}} in order to achieve consistency with \code{\link[dplyr]{lead}} and \code{\link[dplyr]{lag}}.
}
\examples{
returns(1:10)
returns(c(1, 2, 4, 8, 16, 32))

returns(c(1, 2, 4, 8, 16, 32), lag = 2)
}
\seealso{
\code{\link{diff}}
}

