% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scientificLabels}
\alias{scientificLabels}
\title{ggplot labels with exponential notation}
\source{
Brian Diggs \url{https://groups.google.com/forum/#!topic/ggplot2/a_xhMoQyxZ4}
}
\usage{
scientificLabels(l)
}
\arguments{
\item{l}{Labels}
}
\value{
Labels in the desired format
}
\description{
Allows to adjust the numbers display on both axes for ggplot2 plots. The format is such that the expontent is nicely set.
}
\examples{
library(ggplot2)
df <- data.frame(x=rnorm(100), y=rnorm(100))
ggplot(df, aes(x=x, y=y)) +
  geom_point() +
  scale_x_continuous(labels=scientificLabels) +
  scale_y_continuous(labels=scientificLabels)
}

