% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{showCoeftest}
\alias{showCoeftest}
\title{Hides variables in coefficient test}
\usage{
showCoeftest(model, vcov. = NULL, hide = NULL)
}
\arguments{
\item{model}{Object of type \code{lm}.}

\item{vcov.}{A specification of the covariance matrix as used by
\code{\link[lmtest]{coeftest}}.}

\item{hide}{A string. All variables starting with that name are excluded.}
}
\description{
Excludes variables -- usually dummies variables -- following a certain naming scheme. It also appends significance stars.
}
\examples{
x1 <- 1:100
x2 <- rep(c(1, 2), 50)
y <- x1 + x2 + rnorm(100)

m <- lm(y ~ x1 + x2)

showCoeftest(m, hide = "x") # leaves only the intercept
}

