% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{showRowsNA}
\alias{showRowsNA}
\title{Show rows with NA values}
\usage{
showRowsNA(d)
}
\arguments{
\item{d}{A matrix or dataframe.}
}
\value{
Subset of \code{d} containing NA values. Returns \code{NULL} if no NA values
are present.
}
\description{
Show rows that contain an \code{NA} value
}
\examples{
m <- matrix(letters[c(1, 2, NA, 3, NA, 4, 5, 6, 7, 8)],
            ncol = 2, byrow = FALSE)
m
findRowsNA(m)
showRowsNA(m)

d <- data.frame(x = 1:10)
d
findRowsNA(d)
showRowsNA(d)
}
\seealso{
\code{\link{findRowsNA}}
}

