% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{standardizeCoefficients}
\alias{standardizeCoefficients}
\alias{standardizeCoefficients.lm}
\alias{standardizeCoefficients.svarest}
\alias{standardizeCoefficients.varest}
\title{Standardized coefficients}
\usage{
standardizeCoefficients(model, hide = NULL)

\method{standardizeCoefficients}{lm}(model, hide = NULL)

\method{standardizeCoefficients}{varest}(model, hide = NULL)

\method{standardizeCoefficients}{svarest}(model, hide = NULL)
}
\arguments{
\item{model}{Model. Supported types are linear models (\code{lm}) or time series
(\code{varest} or \code{svarest} from the \code{vars} package).}

\item{hide}{A string. All variables starting with that name are excluded.}
}
\value{
Coefficients (with transformations) for model variables. In case of
multivariate models, it returns a list with an element for each dependent variable.
}
\description{
Computes standardized coefficient for a regression or time series. Three scores are
returned:
\itemize{
  \item{"Coef:"} This returns the original coefficient \eqn{\beta}.
  \item{"SdChange"} A one standard deviation in the independent variable
  is linked to percentage increase in the dependent variable. Hence, it is
  \eqn{\beta var(\beta)}.
  \item{"StdCoef:"} Returns the standardized coefficent. It is calculated through
  \eqn{\beta var(\beta) / var(y)} where \eqn{y} is independent variable.
}
}
\examples{
library(dplyr)
library(vars)
data(Canada)

prod <- differences(as.numeric(Canada[, 2]))
production <- data.frame(Prod = prod, Lag1 = dplyr::lag(prod), Lag2 = dplyr::lag(prod, 2))

m <- lm(Prod ~ Lag1, data = production)
standardizeCoefficients(m)

m <- lm(Prod ~ Lag1 + Lag2, data = production)
standardizeCoefficients(m)

var <- VAR(Canada, p = 2, type = "none")
standardizeCoefficients(var$varresult$e)
std <- standardizeCoefficients(var)
std$e
}

