% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{testSpecification}
\alias{testSpecification}
\title{Specification tests for VAR model}
\usage{
testSpecification(var, lags.pt = 15, lags.bg = 5, lags.multi = 5)
}
\arguments{
\item{var}{Estimated vector autoregressive model of type \code{varest} from the
\code{vars} package.}

\item{lags.pt}{An integer specifying the lags to be used for the Portmanteau
statistic.}

\item{lags.bg}{An integer specifying the lags to be used for the Breusch-Godfrey
statistic.}

\item{lags.multi}{An integer specifying the lags to be used for the multivariate
ARCH-LM statistic.}
}
\description{
Tests VAr mode lregarding non-autocorrelation (Breusch-Godfrey, Portmanteau),
normal distribution of residuals (Jarque-Bera) and homoskedasticity (ARCH-LM).
}
\examples{
library(vars)
data(Canada)
var.2c <- VAR(Canada, p = 2, type = "none")

testSpecification(var.2c)
}

