% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{texreg_tvalues}
\alias{texreg_tvalues}
\title{texreg output with t-values}
\usage{
texreg_tvalues(model, hide = NULL, ...)
}
\arguments{
\item{model}{Object of type \code{lm} or a \code{list} of \code{lm} objects.}

\item{hide}{A string. All variables starting with that name are excluded. If \code{NULL} (default),
no variables are omitted.}

\item{...}{Additional parameters that are passed to default \code{\link[texreg]{texreg}} function.}
}
\value{
Object of type \code{\link[texreg]{texregTable}}.
}
\description{
Function is a customized interface to the \code{\link[texreg]{texreg}} function. It replaces
standard errors in the output by t-values.
}
\examples{
x1 <- 1:100
x2 <- rep(c(1, 2), 50)
y <- x1 + x2 + rnorm(100)

m <- lm(y ~ x1 + x2)
texreg_tvalues(m, digits = 4)
}

