% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{correlationMatrix}
\alias{correlationMatrix}
\title{Pretty correlation matrix}
\usage{
correlationMatrix(x, y = NULL, method = c("pearson", "spearman"),
  removeTriangle = c("upper", "lower", "none"), diagonale = FALSE,
  digits = 3, filename = NULL)
}
\arguments{
\item{x}{A numeric matrix for which to compute the correlation matrix.}

\item{y}{Optional second matrix. If omitted or \code{NULL}, it is replaced
by \code{x}.}

\item{method}{Method, i.e. Pearson's \code{r} or Spearman's \code{rho} rank
correlation coefficients.}

\item{removeTriangle}{Parameter indicating whether to remove a duplicated
triangle of the matrix. Defaul is the upper triangle that is removed.}

\item{diagonale}{If \code{FALSE} (default), the diagonale is omitted. If
\code{TRUE}, it is included as 1 along with significance stars.}

\item{digits}{Number of digits to be printed (default: 3).}

\item{filename}{Path to which a LaTeX copy is exported (default is
\code{NULL}). If \code{NULL}, no file is generated.}
}
\value{
Correlation matrix with pretty output.
}
\description{
Calculates a correlation matrix for a dataset and display variables, as well
as P-values in asteriks notation nicely. The function also writes the table
into a LaTeX file if desired.
}
\examples{
data(USArrests)
correlationMatrix(USArrests)

# alternatively, with table exported into a LaTeX file
correlationMatrix(USArrests, filename = "table_cor.tex")
unlink("table_cor.tex")
}

