% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsd}
\alias{cumadev}
\alias{cumkurtosis}
\alias{cumsd}
\alias{cumskewness}
\title{Cumulativate versions of skewness, kurtosis, sd, adv}
\usage{
cumsd(x)

cumskewness(x)

cumkurtosis(x)

cumadev(x)
}
\arguments{
\item{x}{For \code{cumskewness}, \code{cumkurtosis}, \code{cumsd},
\code{cumadv} an integer or numeric vector.}
}
\value{
Returns a vector of the same length with new elements.
}
\description{
Functions include \code{cumskewness}, \code{cumkurtosis},
\code{cumadev} (average deviation), and \code{cumsd} to complete
R's set of cumulate functions to match the aggregation functions available
in most databases.
}
\details{
The most common variants \code{\link{cummin}}, \code{\link{cummax}}
and \code{\link{cumsum}} are included in the \code{"base"} package. In
addition, \code{"dplyr"} ships a \code{\link[dplyr]{cummean}}.
}
\examples{
library(dplyr)

df <- data.frame(x = 1:10, y = rnorm(10))
cumsd(df$x)
df \%>\%
  mutate_all(funs("mean" = cummean, "sd" = cumsd))
}

