% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptiveStatistics}
\alias{descriptiveStatistics}
\title{Pretty summary statistics}
\usage{
descriptiveStatistics(x, digits = 3, subset = c("mean", "median", "min",
  "max", "sd", "skew", "kurtosis"))
}
\arguments{
\item{x}{Dataset in the form of a matrix or a \code{data.frame}.}

\item{digits}{Number of digits to be printed (default: 3).}

\item{subset}{Variables to be selected.}
}
\value{
Matrix with pretty output.
}
\description{
Calculates summary statistics for dataset and display key variables in a nice format.
}
\note{
Class \code{RGT_descriptives} is added to recognize this value inside
\code{\link{export}}.
}
\examples{
data(USArrests)
descriptiveStatistics(USArrests)
}

