% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extract_tvalues}
\alias{extract_tvalues}
\alias{extract_tvalues.lm}
\alias{extract_tvalues.rq}
\title{Extract t-values from lm object}
\usage{
extract_tvalues(model, hide = NULL, se = "nid")

\method{extract_tvalues}{lm}(model, hide = NULL, se = "nid")

\method{extract_tvalues}{rq}(model, hide = NULL, se = "nid")
}
\arguments{
\item{model}{Object of type \code{lm}.}

\item{hide}{A string. All variables starting with that name are excluded. If \code{NULL} (default),
no variables are omitted.}

\item{se}{Computation scheme for standard errors in case of a quantile regression.}
}
\value{
Vector of type \code{numeric}.
}
\description{
Function extracts the t-values from an \code{lm} object as given by the \code{summary} function.
}
\examples{
\dontrun{
x1 <- 1:100
x2 <- rep(c(1, 2), 50)
y <- x1 + x2 + rnorm(100)

m <- lm(y ~ x1 + x2)
extract_tvalues(m)

qr <- rq(y ~ x1 + x2)
extract_tvalues(m)
extract_tvalues(m, se = "ker")
}
}

