% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{impulseResponsePlot}
\alias{impulseResponsePlot}
\title{Calculation and pretty plot of impulse response function}
\usage{
impulseResponsePlot(var, impulse, response, n.ahead = 10, filename = NULL,
  width = 10, height = 6, ...)
}
\arguments{
\item{var}{Object package with impulse response function, i.e. type
\code{varest} or \code{svarest} from the \code{vars}.}

\item{impulse}{String identifier which variable experiences a shock.}

\item{response}{String identfier which variable is the response.}

\item{n.ahead}{Optional parameter to later choose a smaller x-range for the
impulse response function. Argument expects a numeric value with the maximum
step. Default is 10.}

\item{filename}{Filename to export the table as LaTeX. Default is
\code{NULL}, i.e. no export.}

\item{width}{Width of exported figure (default: 10).}

\item{height}{Height of exported figure (default: 6).}

\item{...}{Further arguments passed on to \code{\link{plotIrf}}}
}
\value{
Object of \code{\link[ggplot2]{ggplot}}.
}
\description{
Computes and plots the impulse response function in black/white manner for
suitability with common journals.
}
\examples{
library(vars)
data(Canada)
# For VAR
var.2c <- VAR(Canada, p = 2, type = "const")
\dontrun{
impulseResponsePlot(var.2c, impulse = "e", response = "prod", ylab = "Production")
impulseResponsePlot(var.2c, impulse = "e", response = "prod", ylab = "Production",
                    n.ahead = 5, filename = "irf_e_prod.pdf")
}
}

