% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{remakePackage}
\alias{remakePackage}
\title{Builds, loads and checks package}
\usage{
remakePackage()
}
\description{
Simple wrapper to rebuild, load and check package. This function specifically
updates the manual as well.
}
\examples{
\dontrun{
  remakePackage()
}
}

